Limit socket protocol when detecting peers
When sockets are created with the protocol `ETH_P_ALL`, the receives
can pick up a lot of unnecessary chatter on the line if there are
other broadcasts/multicasts occurring on the network.

Setting the protocol to the EtherType defined by `frame_type` limits
the socket to only receiving frames which match this EtherType.